refactor: SuggetBadeg -> ProposalBadge 네이밍 변경
