+++ src/components/common/badges/Proposal.stories.tsx
@@ -1,0 +1,21 @@
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import ProposalBadge from './ProposalBadge';
+
+export default {
+  title: 'common/badges/ProposalBadge',
+  component: ProposalBadge,
+} as ComponentMeta<typeof ProposalBadge>;
+
+const Template: ComponentStory<typeof ProposalBadge> = (args) => (
+  <ProposalBadge {...args} />
+);
+
+export const 상태_제안중 = Template.bind({});
+상태_제안중.args = {
+  isProposing: true,
+};
+
+export const 상태_거절됨 = Template.bind({});
+상태_거절됨.args = {
+  isProposing: false,
+};
