+++ src/components/home/pending/PendingDongilItem.tsx
@@ -1,4 +1,4 @@
-import SuggestBadge from '@components/common/badges/SuggestBadge';
+import ProposalBadge from '@components/common/badges/ProposalBadge';
 import { modals } from '@components/common/modals/Modals';
 import useModals from '@lib/hooks/useModals';
 import { IDongil } from '@lib/types/IDongil';
@@ -86,8 +86,8 @@ function PendingDongilItem({
           <span className="title">{title}</span>
           <span className="createdAt">{getDate(createdAt)}</span>
         </div>
-        <SuggestBadge
-          isSuggesting={status === EDongilStatus.PENDING ? true : false}
+        <ProposalBadge
+          isProposing={status === EDongilStatus.PENDING ? true : false}
         />
       </StyledButton>
     </>
